Gjenbruke uføresteg for revudering og saksbehandling
